+++ src/plugins/index.js
@@ -1,5 +1,6 @@
 import gitPlugin from './git/index.js';
 import { getPlugin, listPlugins, registerPlugin } from './registry.js';
+import { registerPlugin as registerKernelPlugin } from '../kernel/registries/index.js';
 
 const builtinPlugins = [
   gitPlugin
@@ -8,9 +9,9 @@ const builtinPlugins = [
 export function loadPlugins() {
   for (const plugin of builtinPlugins) {
     registerPlugin(plugin);
+    registerKernelPlugin(plugin);
   }
   return listPlugins();
 }
 
 export { getPlugin, listPlugins };
-